requirements.txt
ipython==6.1.0 # Updated from 6.0.0
django-extensions==1.8.1 # Updated from 1.7.9